requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
msal==1.31.0
requests==2.32.3
openpyxl==3.1.5
pydantic==2.9.2
python-dotenv==1.0.1
